Replace Home container with tsx
