+++ src/containers/Home/Home.tsx
@@ -1,0 +1,36 @@
+import * as tsx from 'vue-tsx-support'
+
+import Presenter, { IPresenter, ListType } from './presenter'
+import PostRepository from '@/repositories/PostRepository'
+import PostList from '@/components/Modules/PostList'
+import styles from './styles.css'
+
+interface IData {
+  listType: ListType
+}
+
+export default tsx.component({
+  name: 'Home',
+  data(): IData {
+    return {
+      listType: ListType.Latest
+    }
+  },
+  computed: {
+    presenter(): IPresenter {
+      return Presenter({
+        postRepository: new PostRepository(this.$store),
+        listType: this.listType
+      })
+    }
+  },
+  render() {
+    return (
+      <div class={styles.Home}>
+        <div class={styles.PostList}>
+          <PostList data={this.presenter.postList} />
+        </div>
+      </div>
+    )
+  }
+})
