+++ src/containers/Home/styles.css
@@ -1,0 +1,22 @@
+.Home {
+  display: flex;
+  flex-direction: column;
+  padding-bottom: 200px;
+}
+
+@media screen and (max-width: 768px) {
+  .Home {
+    width: 100%;
+    padding-top: 32px;
+  }
+}
+
+.PostList {
+  width: 1180px;
+}
+
+@media screen and (max-width: 768px) {
+  .PostList {
+    width: 100%;
+  }
+}
