+++ src/containers/Home/styles.css.d.ts
@@ -1,0 +1,2 @@
+export const Home: string
+export const PostList: string
